requirements.txt
Flask
python-dotenv
PyJWT
requests
pytest
